refactor: move utilities to utils.py
