+++ edx_repo_tools/utils.py
@@ -1,3 +1,7 @@
+import contextlib
+import os
+import subprocess
+
 import click
 from ruamel.yaml import YAML
 
@@ -51,3 +55,32 @@ class YamlLoader:
     def update_yml_file(self):
         with open(self.file_path, 'w') as file_stream:
             self.yml_instance.dump(self.elements, file_stream)
+
+
+def get_cmd_output(cmd):
+    """Run a command in shell, and return the Unicode output."""
+    try:
+        data = subprocess.check_output(cmd, shell=True, stderr=subprocess.STDOUT)
+    except subprocess.CalledProcessError as ex:
+        data = ex.output
+    try:
+        data = data.decode("utf-8")
+    except UnicodeDecodeError:
+        data = data.decode("latin1")
+    return data
+
+
+@contextlib.contextmanager
+def change_dir(new_dir):
+    """Change directory, and then change back.
+
+    Use as a context manager, it will give you the new directory, and later
+    restore the old one.
+
+    """
+    old_dir = os.getcwd()
+    os.chdir(new_dir)
+    try:
+        yield os.getcwd()
+    finally:
+        os.chdir(old_dir)
